Se realizo cambios a la base de datos y se elimino el modulo de galerias

+++ database/migrations/2020_01_27_022137_create_in_shopping_carts_table.php
@@ -14,12 +14,12 @@ class CreateInShoppingCartsTable extends Migration
     public function up()
     {
         Schema::create('in_shopping_carts', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->unsignedBigInteger('shopping_cart_id');
+            $table->increments('id');
+            $table->unsignedInteger('shopping_cart_id');
             $table->foreign('shopping_cart_id')
                 ->references('id')->on('shopping_carts')
                 ->onDelete('cascade');
-            $table->unsignedBigInteger('product_id');
+            $table->unsignedInteger('product_id');
             $table->foreign('product_id')
                 ->references('id')->on('products')
                 ->onDelete('cascade');
